Update oauth package names and fix lint errors
