+++ packages/@magic-ext/react-native-bare-oauth/.eslintrc.js
