+++ packages/@magic-ext/react-native-bare-oauth/.lintstagedrc.yml
