+++ packages/@magic-ext/react-native-bare-oauth/.prettierrc.js
