+++ packages/@magic-ext/react-native-bare-oauth/CHANGELOG.md
