+++ packages/@magic-ext/react-native-bare-oauth/LICENSE
