+++ packages/@magic-ext/react-native-bare-oauth/package.json
