+++ packages/@magic-ext/react-native-bare-oauth/src/crypto.ts
