+++ packages/@magic-ext/react-native-bare-oauth/src/types.ts
