+++ packages/@magic-ext/react-native-bare-oauth/tsconfig.json
