+++ packages/@magic-ext/react-native-expo-oauth/.lintstagedrc.yml
