+++ packages/@magic-ext/react-native-expo-oauth/.prettierrc.js
