+++ packages/@magic-ext/react-native-expo-oauth/LICENSE
