+++ packages/@magic-ext/react-native-expo-oauth/package.json
