+++ packages/@magic-ext/react-native-expo-oauth/src/crypto.ts
