+++ packages/@magic-ext/react-native-expo-oauth/tsconfig.json
